proper unit tests for ContentEmbed (track detection from URL)
